[wasm64] Run wasm worker tests in browser64 (#20550)
Split out from #20516

+++ test/wasm_worker/terminate_all_wasm_workers.c
@@ -52,7 +52,7 @@ int should_throw(void(*func)(emscripten_wasm_worker_t worker), emscripten_wasm_w
     }
     console.error('Function was expected to throw, but did not!');
     return 0;
-  }, (int)func, worker);
+  }, func, worker);
   return threw;
 }
